Bluetooth: Host: Add SMP bt_smp_get_pairing_peer_info
Used to obtain pairing flags from peer before pairing.
Returns bt_smp_pairing for a given bt_conn during smp_pairing_req or
smp_pairing_rsp, which contains IO caps, OOB, AuthReq and keys config.

This function is available through a public SMP API.

Fixes: #21036
