+++ subsys/bluetooth/host/conn.c
@@ -31,7 +31,7 @@
 #include "conn_internal.h"
 #include "l2cap_internal.h"
 #include "keys.h"
-#include "smp.h"
+#include "smp_internal.h"
 #include "att_internal.h"
 #include "gatt_internal.h"
